# schemas/v1/http-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EHAI Execution Plane",
    "version": "2"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/configuration": {
      "get": {
        "summary": "Get Configuration",
        "operationId": "getProjectConfiguration",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "project-configuration.schema.json#/$defs/ProjectConfigurationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update Configuration",
        "operationId": "configureProject",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectConfigurationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Versioned configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "project-configuration.schema.json#/$defs/UpdateProjectConfigurationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/configuration/versions": {
      "get": {
        "summary": "Configuration Versions",
        "operationId": "getProjectConfigurationVersions",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "project-configuration.schema.json#/$defs/ProjectConfigurationVersionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/configuration": {
      "get": {
        "summary": "Run Configuration",
        "operationId": "getRunConfiguration",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "project-configuration.schema.json#/$defs/RunConfigurationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateProjectConfigurationRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "workspace": {
            "type": "string",
            "minLength": 1,
            "title": "Workspace"
          },
          "execution_config_fingerprint": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Config Fingerprint"
          },
          "role_configuration_ref": {
            "type": "string",
            "minLength": 1,
            "title": "Role Configuration Ref"
          },
          "static_rules": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "maxItems": 100,
            "title": "Static Rules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "expected_version",
          "workspace",
          "execution_config_fingerprint",
          "role_configuration_ref",
          "static_rules"
        ],
        "title": "UpdateProjectConfigurationRequest",
        "description": "A complete replacement: empty static_rules clears rules in the new version only."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}